% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonUsage-visualization.R
\docType{methods}
\name{bplot}
\alias{bplot}
\alias{bplot,numeric,numeric,missing-method}
\title{Karlin B plot}
\usage{
\S4method{bplot}{character,character,matrix}(x, y, data,
  annotations = character(), ribosomal = FALSE, reference = list(),
  xlab = character(), ylab = character(), title = character(),
  subtitle = character(), caption = character())

\S4method{bplot}{numeric,numeric,missing}(x, y, data,
  annotations = character(), ribosomal = FALSE, reference = list(),
  xlab = character(), ylab = character(), title = character(),
  subtitle = character(), caption = character())
}
\arguments{
\item{x, y}{Character, both must be in \code{colnames(data)}, or numeric vectors
of CU statistic values for two subsets of genes. If numeric, the vectors
must be of the same length.}

\item{data}{A matrix with CU statistic values for subsets of genes in columns.}

\item{ribosomal}{Logical, whether to indicate ribosomal genes in the plot.
Default is \code{FALSE}, if set to \code{TRUE}, then \code{annotation} must
be given.}

\item{reference}{A named list of logical vector(s) (each of length
\code{nrow(data)}) of reference genes to be indicated on the plot, or a named
list of character vector(s) (of any length) of the reference genes' anotations.
If latter is the case, then \code{annotation} must be given.}

\item{annotation}{A character vector giving KO annotations for sequences
for which the CU values were calculated, must be of length \code{nrow(data)}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distances of each gene's CU frequency to specified gene (sub)sets (given by
\code{x} and \code{y}).
}
