% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCU.R
\name{calcCU}
\alias{calcCU}
\title{Calculation of CU statistics.}
\usage{
calcCU(cdt, method, subsets = list(), self = TRUE, ribosomal = FALSE)
}
\arguments{
\item{cdt}{A codonTable object. Rows represent indivisual sequences. Must
contain the following columns:
\itemize{
  \item ID, a unique sequence identifier
  \item len, sequence length in codons
}
and a single column with counts of every no-stop codon.}

\item{method}{A character string indicating which CU statistic will be
calculated, one of the following: "MILC", "B", "MCB", "ENC", "ENCp", "SCUO".}

\item{subsets}{A (named) list of logical vectors, the length of each equal
to the number of sequences in the set, or codonTable objects of any length.
\code{subsets} should only be supplied if CU statistic that accounts for
background nucleotide composition is calculated, i.e if \code{method} is
one of the following: "MILC", "B", "MCB", "ENCp".}

\item{self}{Logical, if \code{TRUE} (default), CU statistic is also calculated
against the average CU of the entire sequence set. Just as \code{subsets},
this argument is also intended for CU statistics that account for background
nucleotide composition.}

\item{ribosomal}{Logical, if \code{TRUE}, CU statistic is also calculated
against the average CU of the ribosomal genes in the sequence set. This
argument is also intended for CU statistics that account for background
nucleotide composition.}
}
\value{
Returns an input codonTable object augmented with additional columns
  containing CU statistic values for each subset.
}
\description{
Calculates codon usage (CU) statistics for given codonTable object.
}
