% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R
\docType{methods}
\name{codonTable-length}
\alias{codonTable-length}
\title{Length of \code{codonTable} object, i.e. the number
of sequences for which there are codon counts contained in the object.}
\usage{
\S4method{length}{codonTable}(x)
}
\arguments{
\item{x}{A \code{codonTable} object.}
}
\description{
Length of \code{codonTable} object, i.e. the number
of sequences for which there are codon counts contained in the object.
}
