% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R
\docType{methods}
\name{codonTable-show}
\alias{codonTable-show}
\title{Display the object of \code{codonTable} class.}
\usage{
\S4method{show}{codonTable}(object)
}
\arguments{
\item{object}{A \code{codonTable} object.}
}
\description{
Display the object of \code{codonTable} class.
}
