% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRdon.R
\docType{package}
\name{coRdon}
\alias{coRdon}
\alias{coRdon-package}
\title{coRdon: codon usage analysis in R}
\description{
R package for analysis of codone usage in unannotated or
KEGG/COG annotated DNA sequences. Calculates various
measures of CU bias and CU-based predictors of gene
expression, and performs gene set enrichment analysis for
annotated sequences. Implements several methods for
visualization of CU and enrichment analysis results.
}
