% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{class}
\name{crossTab-class}
\alias{crossTab-class}
\alias{getSeqAnnot,crossTab-method}
\alias{getVariable,crossTab-method}
\alias{contable,crossTab-method}
\title{An S4 class \code{crossTab}}
\usage{
\S4method{getSeqAnnot}{crossTab}(x)

\S4method{getVariable}{crossTab}(x)

\S4method{contable}{crossTab}(x)
}
\arguments{
\item{x}{A \code{crossTab} object.}

\item{x}{A \code{crossTab} object.}

\item{x}{A \code{crossTab} object.}
}
\description{
Contingency table of sequences' annotations and the corresponding numeric values.

Get sequence annotations from \code{crossTab} object.

Get values of the variable used to create contingency table
   contained in \code{crossTab} object.

Get contingency table from \code{crossTab} object.
}
\section{Slots}{

\describe{
\item{\code{sequences}}{Character vector of sequences annotations (KO, COG).}

\item{\code{variable}}{Numeric vector of the coresponding CU values.}

\item{\code{table}}{Contingecy table.}
}}

