% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{crossTab-length}
\alias{crossTab-length}
\title{Length of \code{crossTab} object, i.e. the number of
sequences for which the contingency table is contained in the object.}
\usage{
\S4method{length}{crossTab}(x)
}
\arguments{
\item{x}{A \code{crossTab} object.}
}
\description{
Length of \code{crossTab} object, i.e. the number of
sequences for which the contingency table is contained in the object.
}
