% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{crossTab-show}
\alias{crossTab-show}
\title{Display the object of \code{crossTab} class.}
\usage{
\S4method{show}{crossTab}(object)
}
\arguments{
\item{object}{A \code{crossTab} object.}
}
\description{
Display the object of \code{crossTab} class.
}
