% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{crossTab}
\alias{crossTab}
\title{Make a contingency table.}
\usage{
\S4method{crossTab}{character,numeric}(sequences, variable, threshold = 1L,
  percentiles = NULL)
}
\arguments{
\item{sequences}{Character vector of sequences' annotations (KO, COG).}

\item{variable}{Numeric vector of the coresponding CU values.}

\item{threshold}{A threshold value (or a vector of values) of the variable.
Sequences with value of the given variable greater than threshold are
taken as a subset. Default is 1. If no threshold should be set, specify
\code{threshold = NULL}}

\item{percentiles}{A single value or a vector of values between 0 and 1.
Sequences with value of the given variable in the top percentiles are
taken as a subset. If no percentiles should be specified, the argument
takes the value \code{NULL}.}
}
\value{
Returns a data.table with category values in rows, and with separate
   columns for counts in background (all) and subsets, i.e. for diferrent
   thresholds/percentiles provided.
}
\description{
Create a contingency table for the set of annotated sequences and
the corresponding codon usage (CU) values.
}
