% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effNc.R
\name{effNc}
\alias{effNc}
\title{Calculate effective number of codons.}
\usage{
effNc(fa, deg)
}
\arguments{
\item{fa}{a vector containing homozygosity values for all 20 amino acids, or
a matrix with homozygosity values for aminno acids in columns. If \code{fa} is
a matrix, ENC is calculated for every row.}
}
\description{
Calculates effective number of codons (ENC) given set(s) of CU homozygosity
values.
}
