% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich.matrix}
\alias{enrich.matrix}
\title{Extract enrichment values from multiple samples. Note that the samples should
contain annotations of the same type (i.e. the same onthology).}
\usage{
enrich.matrix(x, variable, replace.na = TRUE, ...)
}
\arguments{
\item{x}{list of `enrich.data.frame` objects}

\item{variable}{character string indicating the statistic values to extract
from `enrich.data.frame` objects in x, must be one of `c("enrich","M","A")`.}

\item{replace.na}{logical, whether to replace NA values in the output.
If `TRUE` (default), NAs will be replaced by 0. Alternatively, if numueric,
NAs will be replaced by that given value.}
}
\value{
`matrix` with sequences' annotations as rows, and variable values
  for different samples as columns.
}
\description{
Extract enrichment values from multiple samples. Note that the samples should
contain annotations of the same type (i.e. the same onthology).
}
