% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{enrichment,crossTab-method}
\alias{enrichment,crossTab-method}
\title{Enrichment analysis for codon usage (CU) data.}
\usage{
\S4method{enrichment}{crossTab}(x, pvalueCutoff = numeric(),
  pAdjustMethod = "BH", padjCutoff = numeric())
}
\arguments{
\item{x}{`crossTab` object}

\item{pAdjustMethod}{character, one of the `p.adjust.methods`.}

\item{padjCutoff}{numeric, discard categories with adjusted p value below
this threshold. By default, no threshold is set (`numeric()`).}

\item{pvaluCutoff}{numeric, discard categories with p value below this
threshold. By default, no threshold is set (`numeric()`).}
}
\value{
An `enrich.data.frame` object, or a list of those.
}
\description{
Performs enrichment analysis, given a contongency table of codon counts.
p values are calculated by binomial test, adjustment for multiple testing
can be performed by any of the `p.adjust.methods`.
}
