% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{enrichment}
\alias{enrichment}
\alias{enrichment,crossTab-method}
\title{Enrichment analysis for object \code{crossTab} class.}
\usage{
enrichment(x, ...)

\S4method{enrichment}{crossTab}(x, pvalueCutoff = numeric(),
  pAdjustMethod = "BH")
}
\arguments{
\item{x}{A \code{crossTab} object.}

\item{target}{A character vector indicating which onthology to use, either
\code{"pathway"} or \code{"module"}.}
}
\value{
Returns a list of data.tables, each containing categories in rows,
   and the following columns:
   \itemize{
     \item background counts
     \item counts for a given subset
     \item enrichment, calculated as the ratio:
         (scaled sample counts - scaled backg. counts) / scaled backg. counts * 100,
         where scaling means that sample counts are simply increased by 1,
         and background counts are multiplied by ratio of summed sample counts
         and summed backgroun counts, and also increased by 1.
     \item M, log ratios of scaled counts
     \item A, mean average of scaled counts
     \item pvals, p values for exact binomial test
     \item padj, p values corrected by BH method.
   }
}
\description{
Enrichment analysis for object \code{crossTab} class.
}
\section{Methods (by class)}{
\itemize{
\item \code{crossTab}: Enrichment analysis.
}}

