% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectCU.R
\name{expectCU}
\alias{expectCU}
\title{Calculation of expected codon frequencies.}
\usage{
expectCU(cdt, ctab, subsets, self, ribosomal)
}
\arguments{
\item{cdt}{A codonTable object. Rows represent indivisual sequences. Must
contain the following columns:
\itemize{
  \item ID, a unique sequence identifier
  \item len, sequence length in codons
}
and a single column with counts of every no-stop codon.}

\item{subsets}{A (named) list of logical vectors, the length of each equal
to the number of sequences in the set, or character vectors (of any length)
containing KEGG/eggNOG annotations, or codonTable objects (of any length).
\code{subsets} should only be supplied if CU statistic that accounts for
background nucleotide composition is calculated, i.e if \code{method} is
one of the following: "MILC", "B", "MCB", "ENCp". Otherwise \code{subsets}
parameter is ignored.}

\item{self}{Logical, if \code{TRUE} (default), CU statistic is also calculated
against the average CU of the entire sequence set. Just as \code{subsets},
this argument is also intended for CU statistics that account for background
nucleotide composition.}

\item{ribosomal}{Logical, if \code{TRUE}, CU statistic is also calculated
against the average CU of the ribosomal genes in the sequence set. This
argument is also intended for CU statistics that account for background
nucleotide composition.}
}
\value{
A 61 x \code{length(subsets)} matrix with alphabetically ordered codons'
  frequencies (rows) for every subset (column).
}
\description{
Calculates expected codon frequencies based on CU distribution for each
\code{subsets} of sequences in \code{cdt}.
}
