% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonUsage-expressivity.R
\docType{methods}
\name{GCB}
\alias{GCB}
\alias{GCB,codonTable-method}
\title{Calculate gene codon bias (GCB).}
\usage{
GCB(cTobject, seed = logical(), ribosomal = FALSE, perc = 0.05,
  id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE,
  filtering = "none", len.threshold = 80)

\S4method{GCB}{codonTable}(cTobject, seed = logical(), ribosomal = FALSE,
  perc = 0.05, id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE,
  filtering = "none", len.threshold = 80)
}
\arguments{
\item{cTobject}{A \code{codonTable} object.}

\item{seed}{A logical vector, of the length equal to
\code{getlen(cTobject)}, or a character vector (of any length)
containing KEGG/eggNOG annotations, or a codonTable object
(of any length). Indicates a set of genes, or their CU, to be used
as a target in the first iteration of the algorithm.}

\item{ribosomal}{Logical, if \code{TRUE}, ribosomal genes are used
as a seed. Default is \code{FALSE}, in which case \code{seed}
has to be specified.}

\item{perc}{percent of top ranking genes to be used as a target set
for the next iteration. Default is 0.05.}

\item{id_or_name2}{
    A single string that uniquely identifies the genetic code to extract.
    Should be one of the values in the \code{id} or \code{name2} columns
    of \code{GENETIC_CODE_TABLE}.
  }

\item{alt.init}{logical, whether to use alternative initiation codons.
Default is \code{TRUE}.}

\item{stop.rm}{Logical, whether to remove stop codons. Default is
\code{FALSE}.}

\item{filtering}{Character vector, one of \code{c("none", "soft", "hard")}.
Specifies whether sequences shorther than some threshold value of
length (in codons), \code{len.threshold}, should be excluded from
calculations. If \code{"none"} (default), length of sequences is not
checked, if \code{"soft"}, a warrning is printed if there are shorter
sequences, and if \code{"hard"}, these sequences are excluded from
calculation.}

\item{len.threshold}{Optional numeric, specifying sequence length,
in codons, used for filtering.}
}
\value{
A numeric vector with GCB values.

   For definition of gene codon bias (GCB), see
   \href{https://www.ncbi.nlm.nih.gov/pubmed/6175758}{Ikemura (1981)}.
}
\description{
Calculate GCB values for every sequence in the given
\code{codonTable} object.
}
\examples{
# load example DNA sequences
exampledir <- system.file("extdata", package = "coRdon")
cT <- codonTable(readSet(exampledir))

# calculate GCB with CU of ribosomal genes among the example DNA sequences
# used as a target (seed) in the first iteration of the algorithm
gcb <- GCB(cT, ribosomal = TRUE)
head(gcb)

# calculate GCB distance with the first 20 example DNA sequences
# (i.e. the first half of the example DNA set) as a seed
gcb <- GCB(cT, seed = c(rep(TRUE, 20), rep(FALSE, 20)))

# alternatively, you can specify codonTable as a seed
halfcT <- codonTable(codonCounts(cT)[1:20,])
gcb2 <- GCB(cT, seed = halfcT)
all.equal(gcb, gcb2) # TRUE

# options for genetic code
gcb <- GCB(cT, ribosomal = TRUE,
           stop.rm = TRUE) # don't use stop codons in calculation
gcb <- GCB(cT, ribosomal = TRUE,
           alt.init = FALSE) # don't use alternative start codons
gcb <- GCB(cT, ribosomal = TRUE,
           id_or_name2 = "2") # use different genetic code, for help
                              # see `?Biostrings::GENETIC_CODE`

}
