% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HD59}
\alias{HD59}
\title{Codon usage in healthy human gut microbiome.}
\format{A \code{codonTable} object.}
\source{
\href{https://go.nature.com/2G7QdpC}{Quin et al. 2014};
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}
}
\usage{
HD59
}
\description{
A \code{codonTable} object with codon counts for sequences of the human gut
metagenome, from a healthy individual. Raw sequences are from
\href{https://go.nature.com/2G7QdpC}{Quin et al. 2014},
processed, assembled and annotated (KEGG Orthology) as descrbed in
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}.
}
\keyword{datasets}
