% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HD59_KO}
\alias{HD59_KO}
\title{Codon usage based KO enrichment analysis results
from the healthy human gut microbiome.
For more information, see `?HD59`.}
\format{An \code{AnnotatedDataFrame} object.
See `?enrichment` for description.}
\source{
\href{https://go.nature.com/2G7QdpC}{Quin et al. 2014};
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}
}
\usage{
HD59_KO
}
\description{
Codon usage based KO enrichment analysis results
from the healthy human gut microbiome.
For more information, see `?HD59`.
}
\keyword{datasets}
