% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HD59_PATHWAYS}
\alias{HD59_PATHWAYS}
\title{Codon usage based KEGG Pathway enrichment analysis results
from a healthy human gut microbiome.}
\format{An \code{AnnotatedDataFrame} object.
See `?enrichment` for description.}
\source{
\href{https://go.nature.com/2G7QdpC}{Quin et al. 2014};
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}
}
\usage{
HD59_PATHWAYS
}
\description{
Codon usage based KEGG Pathway enrichment analysis results
from a healthy human gut microbiome.
}
\keyword{datasets}
