% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LD94_KO}
\alias{LD94_KO}
\title{Codon usage based KO enrichment analysis results
from an gut microbiome of an individual with liver cirrhosis.
For more information, see `?LD94`.}
\format{An \code{AnnotatedDataFrame} object.
See `?enrichment` for description.}
\source{
\href{https://go.nature.com/2G7QdpC}{Quin et al. 2014};
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}
}
\usage{
LD94_KO
}
\description{
Codon usage based KO enrichment analysis results
from an gut microbiome of an individual with liver cirrhosis.
For more information, see `?LD94`.
}
\keyword{datasets}
