% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R
\docType{methods}
\name{length-codonTable}
\alias{length-codonTable}
\alias{length,codonTable-method}
\title{Length of \code{codonTable} object.}
\usage{
\S4method{length}{codonTable}(x)
}
\arguments{
\item{x}{A \code{codonTable} object.}
}
\value{
Numeric, the length of \code{x}.
}
\description{
Length of \code{codonTable} object is the number of sequences
for which there are codon counts contained in the object.
}
