% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R, R/crossTab-class.R
\docType{methods}
\name{length,codonTable-method}
\alias{length,codonTable-method}
\alias{length,crossTab-method}
\title{Length of \code{codonTable} object}
\usage{
\S4method{length}{codonTable}(x)

\S4method{length}{crossTab}(x)
}
\arguments{
\item{x}{A \code{codonTable} object.}

\item{x}{A \code{crossTab} object.}
}
\description{
Get the length of \code{codonTable} object, i.e. the number of sequences
for which there are codon counts contained in the object.

Get the length of \code{crossTab} object, i.e. the number of sequences
for which the contingency table is contained in the object.
}
