% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{make.contable}
\alias{make.contable}
\title{Make contingency table}
\usage{
make.contable(geneList, variable, threshold = 1L, percentiles = NULL)
}
\arguments{
\item{geneList}{A character vector of sequences annotations (KO, COG).}

\item{variable}{A coresponding numeric vector of CU values.}

\item{threshold}{A threshold value (or a vector of values) of the variable.
Sequences with value of the given variable greater than threshold are
counted as a subset. Default is 1.}

\item{percentiles}{A single value or a vector of values between 0 and 1.
Sequences with value of the given variable in the top percentiles are
counted as a subset.}
}
\value{
Returns a data.table with category values in rows, and with separate
   columns for counts in background (all) and subsets, i.e. for diferrent
   thresholds/percentiles provided.
}
\description{
Creates a contingency table for the list of annotated sequences and
the corresponding list of codon usage (CU) values.
}
