% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonUsage.R
\docType{methods}
\name{MILC}
\alias{MILC}
\title{Calculate codon usage Measure Independent of Length and Composition (MILC)}
\usage{
\S4method{MILC}{codonTable}(cTobject, subsets = list(), self = TRUE,
  ribosomal = FALSE, id_or_name2 = "1", alt.init = TRUE,
  stop.rm = FALSE, filtering = "none", len.threshold = 80)
}
\arguments{
\item{cTobject}{A \code{codonTable} object.}

\item{subsets}{A (named) list of logical vectors, the length of each equal
to \code{getlen(cTobject)}, i.e. the number of sequences in the set, or
character vectors (of any length) containing KEGG/eggNOG annotations,
or codonTable objects (of any length).}

\item{self}{Logical, if \code{TRUE} (default), CU statistic is also calculated
against the average CU of the entire set of sequences.}

\item{ribosomal}{Logical, if \code{TRUE}, CU statistic is also calculated
against the average CU of the ribosomal genes in the sequence set.}

\item{id_or_name2}{
    A single string that uniquely identifies the genetic code to extract.
    Should be one of the values in the \code{id} or \code{name2} columns
    of \code{GENETIC_CODE_TABLE}.
  }

\item{alt.init}{logical, whether to use alternative initiation codons.
Default is \code{TRUE}.}

\item{stop.rm}{Logical, whether to remove stop codons. Default is \code{FALSE}.}

\item{filtering}{Character vector, one of \code{c("none", "soft", "hard")}.
Specifies wether sequences shorther than some threshold value of length
(in codons), \code{len.threshold}, should be excluded from calculations. If
\code{"none"} (default), length of sequences is not checked, if \code{"soft"},
a warrning is printed if there are shorter sequences, and if \code{"hard"},
these sequences are excluded.}

\item{len.threshold}{Optional numeric, specifying sequence length, in codons,
used for filtering.}
}
\value{
A matrix with MILC values for every specified subset (\code{subsets},
   \code{self}, \code{ribosomal}) in columns.

   For definition of MILC, see
   \href{https://www.ncbi.nlm.nih.gov/pubmed/16029499}{Supek & Vlahovicek (2005)}.
}
\description{
Calculate MILC values for every sequence in the given \code{codonTable} object.
}
