% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSet.R
\name{readSet}
\alias{readSet}
\title{Read Set}
\usage{
readSet(folder = ".", KOs = c(), zipped = FALSE)
}
\value{
Returns codon usage table for all files.
}
\description{
Reads sets of fasta sequences stored in \code{folder}.
}
