% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{reduce.contable}
\alias{reduce.contable}
\title{Reduce contingency table}
\usage{
reduce.contable(contable, target)
}
\arguments{
\item{contable}{A contingency table.}

\item{target}{A character vector indicating which onthology to use, either
\code{"pathway"} or \code{"module"}.}
}
\value{
Returns a data.table with new category values in rows, and with the same
   columns as in the input contingency table.
}
\description{
Given contingency table with gene identifiers, reduce the table by associating
gene identifiers with either KEGG Pathway or KEGG Module identifiers.
}
