% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{reduceCrossTab}
\alias{reduceCrossTab}
\alias{reduceCrossTab,crossTab,character-method}
\title{Reduce objects of \code{crossTab} class.}
\usage{
reduceCrossTab(x, target)

\S4method{reduceCrossTab}{crossTab,character}(x, target)
}
\arguments{
\item{x}{A \code{crossTab} object to be reduced.}

\item{target}{A character vector indicating which onthology to use, either
\code{"pathway"} or \code{"module"}.}
}
\value{
Returns input \code{crossTab} object, with updated contingency table,
   displaying new category values in rows, and updated counts in columns.
}
\description{
Reduce objects of \code{crossTab} class.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = crossTab,target = character}: Reduce \code{crossTab}.

Reduce the input contingency table by associating genes with KEGG Pathway
or KEGG Module identifiers.
}}

