% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RPKOs}
\alias{RPKOs}
\title{KEGG Orthology (KO) annotations for ribosomal genes.}
\format{A character vector.}
\usage{
RPKOs
}
\description{
KEGG Orthology (KO) annotations for ribosomal genes.
}
\keyword{datasets}
