% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R, R/crossTab-class.R
\docType{methods}
\name{show,codonTable-method}
\alias{show,codonTable-method}
\alias{show,crossTab-method}
\title{Display the object of \code{codonTable} class.}
\usage{
\S4method{show}{codonTable}(object)

\S4method{show}{crossTab}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Display the object of \code{codonTable} class.

Display the object of \code{crossTab} class.
}
