% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonUsage.R
\docType{methods}
\name{ENCprime}
\alias{ENCprime}
\alias{ENCprime,codonTable-method}
\title{Calculate effective number of codons prime (ENC').}
\usage{
ENCprime(cTobject, subsets = list(), self = TRUE, ribosomal = FALSE,
  id_or_name2 = "1", alt.init = TRUE, stop.rm = TRUE,
  filtering = "none", len.threshold = 80)

\S4method{ENCprime}{codonTable}(cTobject, subsets = list(), self = TRUE,
  ribosomal = FALSE, id_or_name2 = "1", alt.init = TRUE, stop.rm = TRUE,
  filtering = "none", len.threshold = 80)
}
\arguments{
\item{cTobject}{A \code{codonTable} object.}

\item{subsets}{A (named) list of logical vectors, the length of each equal
to \code{getlen(cTobject)}, i.e. the number of sequences in the set, or
character vectors (of any length) containing KEGG/eggNOG annotations,
or codonTable objects (of any length).}

\item{self}{Logical, if \code{TRUE} (default), CU statistic is also
calculated against the average CU of the entire set of sequences.}

\item{ribosomal}{Logical, if \code{TRUE}, CU statistic is also calculated
against the average CU of the ribosomal genes in the sequence set.}

\item{id_or_name2}{
    A single string that uniquely identifies the genetic code to extract.
    Should be one of the values in the \code{id} or \code{name2} columns
    of \code{GENETIC_CODE_TABLE}.
  }

\item{alt.init}{logical, whether to use alternative initiation codons.
Default is \code{TRUE}.}

\item{stop.rm}{Logical, whether to remove stop codons. Default is
\code{FALSE}.}

\item{filtering}{Character vector, one of \code{c("none", "soft", "hard")}.
Specifies whether sequences shorther than some threshold value of
length (in codons), \code{len.threshold}, should be excluded from
calculations. If \code{"none"} (default), length of sequences is not
checked, if \code{"soft"}, a warrning is printed if there are shorter
sequences, and if \code{"hard"}, these sequences are excluded from
calculation.}

\item{len.threshold}{Optional numeric, specifying sequence length,
in codons, used for filtering.}
}
\value{
A matrix with ENC' values for every specified subset
   (\code{subsets}, \code{self}, \code{ribosomal}) in columns.

   For definition of effective number of codons prime (ENC'), see
   \href{https://www.ncbi.nlm.nih.gov/pubmed/12140252}{Novembre (2002)}.
}
\description{
Calculate ENC' values for every sequence in the given
\code{codonTable} object.
}
\examples{
# load example DNA sequences
exampledir <- system.file("extdata", package = "coRdon")
cT <- codonTable(readSet(exampledir))

# calculate ENC' distance to the average CU of the example DNA sequences
encp <- ENCprime(cT)
head(encp)

# also calculate ENC' distance to the average CU
# of ribosomal genes among the example DNA sequences
encp <- ENCprime(cT, ribosomal = TRUE)
head(encp)

# calculate ENC' distance to the average CU
# of the first 20 example DNA sequences
# (i.e. the first half of the example DNA set)
encp <- ENCprime(cT, self = FALSE,
                 subsets = list(half = c(rep(TRUE, 20), rep(FALSE, 20))))

# alternatively, you can specify codonTable as a subset
halfcT <- codonTable(codonCounts(cT)[1:20,])
encp2 <- ENCprime(cT, self = FALSE, subsets = list(half = halfcT))
all.equal(encp, encp2) # TRUE

# filtering
ENCprime(cT, filtering = "hard", len.threshold = 80) # ENC' for 9 sequences
sum(getlen(cT) > 80) # 9 sequences are longer than 80 codons
encp1 <- ENCprime(cT, filtering = "none") # no filtering
encp2 <- ENCprime(cT, filtering = "soft") # warning
all.equal(encp1, encp2) # TRUE

# options for genetic code
encp <- ENCprime(cT, stop.rm = TRUE) # don't use stop codons in calculation
encp <- ENCprime(cT, alt.init = FALSE) # don't use alternative start codons
encp <- ENCprime(cT, id_or_name2 = "2") # use different genetic code,
                                        # see `?Biostrings::GENETIC_CODE`

}
