% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{enrich.matrix}
\alias{enrich.matrix}
\title{Extract chosen enrichment values to a matrix.}
\usage{
\S4method{enrich.matrix}{list}(x, variable, replace.na)
}
\arguments{
\item{x}{list of \code{enrich.data.frame} objects}

\item{variable}{Character, indicating the statistic values to extract from
\code{enrich.data.frame} objects in x, must be one of \code{c("enrich","M","A")}.}

\item{replace.na}{logical, whether to replace NA values in the output.
If `TRUE` (default), NAs will be replaced by 0. Alternatively, if numueric,
NAs will be replaced by that given value.}
}
\value{
\code{matrix} with sequences' annotations as rows, and variable values
  for different samples as columns.
}
\description{
Extract enrichment values from multiple samples, i.e. \code{enrich.data.frame}
objects. Note that the samples should contain annotations of the same type
(i.e. the same ontology). The data in matrix format can be efortlessly used in
different types downstream analyses, such as GAGE, and visualised, e.g. using
a heatmap.
}
