% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-visualization.R
\docType{methods}
\name{enrichBarplot}
\alias{enrichBarplot}
\alias{enrichBarplot,list-method}
\alias{enrichBarplot,enrich.data.frame-method}
\title{Barplot of enriched and depleted annotations.}
\usage{
enrichBarplot(x, variable, pvalue = "pvals", alpha = numeric())

\S4method{enrichBarplot}{list}(x, variable, pvalue = "pvals",
  alpha = numeric())

\S4method{enrichBarplot}{enrich.data.frame}(x, variable, pvalue = "pvals",
  alpha = numeric())
}
\arguments{
\item{x}{\code{enrich.data.frame} object, or a list of those.}

\item{variable}{Character, indicating the statistic values to be used for
plotting, must be one of \code{c("enrich","M","A")}.}

\item{pvalue}{Character, one of \code{c("pvals", "padj")}.}

\item{alpha}{Numeric, significance level to be used for plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
Make a barplot of enriched annotations. Bars' heights represent values of
the chosen enrichment statistic (\code{c("enrich","M","A")}), and the colours
represent the p values (\code{c("pvals", "padj")}).
}
