% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-visualization.R
\docType{methods}
\name{enrichMAplot}
\alias{enrichMAplot}
\alias{enrichMAplot,list-method}
\alias{enrichMAplot,AnnotatedDataFrame-method}
\title{MA plot of enriched annotations.}
\usage{
enrichMAplot(x, pvalue = "pvals", alpha = 0.05)

\S4method{enrichMAplot}{list}(x, pvalue = "pvals", alpha = 0.05)

\S4method{enrichMAplot}{AnnotatedDataFrame}(x, pvalue = "pvals",
  alpha = 0.05)
}
\arguments{
\item{x}{\code{AnnotatedDataFrame} object, or a list of those.}

\item{pvalue}{Character, one of \code{c("pvals", "padj")}.}

\item{alpha}{Numeric, significance level to be used for plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
Make an MA-like plot of enriched annotations, similar to the commonly used
plots in differential expression analysis.
}
