% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-visualization.R
\docType{methods}
\name{enrichMAplot}
\alias{enrichMAplot}
\alias{enrichMAplot,enrich.data.frame-method}
\title{MA plot of enriched annotations.}
\usage{
\S4method{enrichMAplot}{list}(x, pvalue, alpha)

\S4method{enrichMAplot}{enrich.data.frame}(x, pvalue, alpha)
}
\arguments{
\item{x}{\code{enrich.data.frame} object, or a list of those.}

\item{pvalue}{Character, one of \code{c("pvals", "padj")}.}

\item{alpha}{Numeric, significance level to be used for plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
Make an MA-like plot of enriched annotations, similar to the commonly used
plots in differential expression analysis.
}
