% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{enrichment}
\alias{enrichment}
\title{Enrichment analysis for codon usage (CU) data.}
\usage{
\S4method{enrichment}{crossTab}(x, pvalueCutoff = numeric(),
  pAdjustMethod = "BH", padjCutoff = numeric())
}
\arguments{
\item{x}{A \code{crossTab} object}

\item{pAdjustMethod}{Character, one of the \code{p.adjust.methods}.}

\item{padjCutoff}{Numeric, discard categories with adjusted p value below
this threshold. By default, no threshold is set (\code{numeric()}).}

\item{pvaluCutoff}{Numeric, discard categories with p value below this
threshold. By default, no threshold is set (\code{numeric()}).}
}
\value{
An \code{enrich.data.frame} object, or a list of those.
}
\description{
Performs enrichment analysis, given a contongency table of codon counts.
p values are calculated by binomial test, adjustment for multiple testing
can be performed by any of the \code{p.adjust.methods}.
}
