% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonUsage-expressivity.R
\docType{methods}
\name{GCB}
\alias{GCB}
\title{Calculate gene codon bias (GCB).}
\usage{
\S4method{GCB}{codonTable}(cTobject, seed, ribosomal = FALSE, perc = 0.05,
  id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE,
  filtering = "none", len.threshold = 80)
}
\arguments{
\item{cTobject}{A \code{codonTable} object.}

\item{seed}{A logical vector, of the length equal to \code{getlen(cTobject)},
or a character vector (of any length) containing KEGG/eggNOG annotations,
or a codonTable object (of any length). Indicates a set of genes, or their CU,
to be used as a target in the first iteration of the algorithm.}

\item{ribosomal}{Logical, if \code{TRUE}, ribosomal genes are used as a seed.}

\item{perc}{percent of top ranking genes to be used as a target set for the
next iteration}

\item{id_or_name2}{
    A single string that uniquely identifies the genetic code to extract.
    Should be one of the values in the \code{id} or \code{name2} columns
    of \code{GENETIC_CODE_TABLE}.
  }

\item{alt.init}{logical, whether to use alternative initiation codons.
Default is \code{TRUE}.}

\item{stop.rm}{Logical, whether to remove stop codons. Default is \code{FALSE}.}

\item{filtering}{Character vector, one of \code{c("none", "soft", "hard")}.
Specifies wether sequences shorther than some threshold value of length
(in codons), \code{len.threshold}, should be excluded from calculations. If
\code{"none"} (default), length of sequences is not checked, if \code{"soft"},
a warrning is printed if there are shorter sequences, and if \code{"hard"},
these sequences are excluded.}

\item{len.threshold}{Optional numeric, specifying sequence length, in codons,
used for filtering.}
}
\value{
A numeric vector with GCB values.

   For definition of gene codon bias (GCB), see
   \href{https://www.ncbi.nlm.nih.gov/pubmed/6175758}{Ikemura (1981)}.
}
\description{
Calculate GCB values for every sequence in the given \code{codonTable} object.
}
