% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LD94_PATHWAYS}
\alias{LD94_PATHWAYS}
\title{Codon usage based  KEGG Pathway enrichment analysis results
from an gut microbiome of an individual with liver cirrhosis.}
\format{An \code{AnnotatedDataFrame} object.
See `?enrichment` for description.}
\source{
\href{https://www.nature.com/articles/nature13568}{Quin et al. 2014};
\href{https://bit.ly/2DRfiz6}{Fabijanic and Vlahovicek 2016}
}
\usage{
LD94_PATHWAYS
}
\description{
Codon usage based  KEGG Pathway enrichment analysis results
from an gut microbiome of an individual with liver cirrhosis.
}
\keyword{datasets}
