% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{length-crossTab}
\alias{length-crossTab}
\alias{length,crossTab-method}
\title{Length of \code{crossTab} object.}
\usage{
\S4method{length}{crossTab}(x)
}
\arguments{
\item{x}{A \code{crossTab} object.}
}
\value{
Numeric, the length of \code{x}.
}
\description{
The length of \code{crossTab} is number of sequences for which
the contingency table is contained in the object.
}
