% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{reduceCrossTab}
\alias{reduceCrossTab}
\title{Reduce \code{crossTab}.}
\usage{
\S4method{reduceCrossTab}{crossTab,character}(x, target)
}
\arguments{
\item{x}{\code{crossTab} object to be reduced.}

\item{target}{Character vector indicating which onthology to use, either
\code{"pathway"} or \code{"module"}.}
}
\value{
Returns input \code{crossTab} object, with updated contingency table,
   displaying new category values in rows, and updated counts in columns.
}
\description{
Reduce the input contingency table by associating sequences with KEGG Pathway
or KEGG Module identifiers.
}
