% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab-class.R
\docType{methods}
\name{show-crossTab}
\alias{show-crossTab}
\alias{show,crossTab-method}
\title{Display the object of \code{crossTab} class.}
\usage{
\S4method{show}{crossTab}(object)
}
\arguments{
\item{object}{A \code{crossTab} object.}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Display the object of \code{crossTab} class.
}
