% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,codonTable,logical-method}
\alias{subset,codonTable,character-method}
\title{Subsetting objects of \code{codonTable} class.}
\usage{
subset(x, subset)

\S4method{subset}{codonTable,logical}(x, subset)

\S4method{subset}{codonTable,character}(x, subset)
}
\arguments{
\item{x}{A \code{codonTable} object to be subset.}

\item{subset}{A logical or character vector indicating which elements of
\code{cTobject} to keep. If logical, \code{subset} should be of length
\code{nrow(counts(cTobject))}. If character, \code{subset} should contain
 at least some of the elements of either \code{getKO(codonTable)} or
 \code{getCOG(codonTable)}.}
}
\description{
Return subsets of \code{codonTable} object, keeping in each slot only those
elements that meet the criteria in \code{subset}.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = codonTable,subset = logical}: Subset \code{codonTable} object.
}}

