% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonTable-class.R
\docType{methods}
\name{subset,codonTable-method}
\alias{subset,codonTable-method}
\title{Subset \code{codonTable} object.}
\usage{
\S4method{subset}{codonTable}(x, subset)
}
\arguments{
\item{x}{A \code{codonTable} object to be subset.}

\item{subset}{A logical or character vector indicating which elements of
\code{cTobject} to keep. If logical, \code{subset} should be of length
\code{nrow(counts(cTobject))}. If character, \code{subset} should contain
 at least some of the elements of either \code{getKO(codonTable)} or
 \code{getCOG(codonTable)}.}
}
\value{
subsets of \code{codonTable} object, keeping in each slot only those
elements that meet the criteria in \code{subset}.
}
\description{
Subset \code{codonTable} object.
}
